new instagrapi/requirements.txt
flask
python-dotenv
instagrapi
google-generativeai
